% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarparkAvailability.R
\name{carpark_availability}
\alias{carpark_availability}
\title{Carpark Availability}
\usage{
carpark_availability(date = "")
}
\arguments{
\item{date}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing the carpark id, type, last update, total lots, and current lots.
}
\description{
This functions calls upon the carpark availability API from data.gov.sg
and processes the returning page.
}
\examples{
dgsg_carpark()
dgsg_carpark(date = "2019-06-05T10:10:10")
dgsg_carpark(date = "2018-07-01T19:32:56")
}
\keyword{carpark}
