% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carpark_info.R
\name{get_carpark_info}
\alias{get_carpark_info}
\title{Carpark Info}
\usage{
get_carpark_info()
}
\value{
A dataframe containing the carpark information.
Extracted fields include car_park_no, address, x_coord, y_coord, car_park_type,
type_of_parking_system, short_term_parking, free_parking,
car_park_decks, gantry_height, car_park_basement
}
\description{
This functions calls upon the datastore_search API from data.gov.sg
to extract HDB carpark information and return it as data.frame
}
\examples{
get_carpark_info()

carpark_availability() \%>\%
     left_join(get_carpark_info(), by = c("id" = "car_park_no")) \%>\%
     as_tibble()
}
