% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseApiDate.R
\name{parse_api_date}
\alias{parse_api_date}
\title{ParseApiDate}
\usage{
parse_api_date(api, input_date = "", summary)
}
\arguments{
\item{api}{The api to be called, i.e. api = "environment/air-temperature"}

\item{input_date}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS or YYYY-MM-DD}

\item{summary}{Returns an error if the user inputs a date only time when not supported by the API.}
}
\value{
A url to be called by the API.
}
\description{
Helper function to check if the date was input correctly
before returning a url to be called by the GET function.
Also returns an error if the API called can only provide data
for a specific date-time, and not date only.
}
