% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSI.R
\name{psi}
\alias{psi}
\title{PSI(PoullutantStandardIndex)}
\usage{
psi(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing various PSI measures across 5 corners
of Singapore
}
\description{
This functions calls upon the PSI API from data.gov.sg
and returns a data frame of the different measures of the PSI across 5
different areas in Singapores and the overall measure for the given
data-time. This data provided by the API is updated hourly.
}
\details{
Note that this function is different from the `PSI_summary` function,
which returns the PSI measures for a given day.
}
\examples{
psi()
psi(date = "2019-11-08T17:30:00")
psi(date = "2018-01-04T09:16:17")
}
\keyword{PSI}
