% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSISummary.R
\name{psi_summary}
\alias{psi_summary}
\title{PSI(PoullutantStandardIndex) Summary}
\usage{
psi_summary(date = "")
}
\arguments{
\item{date}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing various PSI measures across 5 corners
of Singapore and time of day. Dependent on the data availible, not all
results range from 0000 to 2300.
}
\description{
This functions calls upon the PSI API from data.gov.sg
and returns a data frame of the different measures of the PSI across 5
different areas in Singapores and the overall measure for the given
date. This data provided by the API is updated hourly.
}
\details{
Note that this function is different from the `PSI` function,
which returns the PSI measures for a given date-time.
}
\examples{
psi_summary()
psi_summary(date = "2019-11-08")
psi_summary(date = "2018-01-04")
}
\keyword{PSI}
