% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiAvailability.R
\name{taxi_availability}
\alias{taxi_availability}
\title{Taxi Availability}
\usage{
taxi_availability(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing the longitude and latitude values of availible taxis.
}
\description{
This functions calls upon the taxi availability API from data.gov.sg
and returns a data frame of the locations of all availible taxis.
}
\examples{
taxi_availability()
taxi_availability(date = "2019-08-07T09:30:00")
taxi_availability(date = "2018-06-05T13:45:00")
}
\keyword{taxi}
