% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrafficImages.R
\name{traffic_images}
\alias{traffic_images}
\title{Traffic Images}
\usage{
traffic_images(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing links to the traffic images of current
traffic conditions
}
\description{
This functions calls upon the Traffic Images API from data.gov.sg
and returns links to images of live traffic conditions along expressways and
Woodlands & Tuas Checkpoints, including details of the camera location.
This data provided by the API is updated every minute.
}
\examples{
traffic_images()
traffic_images(date = "2022-07-01T15:32:45")
traffic_images(date = "2021-02-11T14:11:07")
}
\keyword{Traffic}
