% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVI.R
\name{uvi}
\alias{uvi}
\title{UVI(Ultra-violet Index)}
\usage{
uvi(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing the current and past hourly UVI
}
\description{
This functions calls upon the UVI API from data.gov.sg
and returns a data frame of the different measures of the UVI across
Singapore and returns the closest UVI value presently and for the past
few hours. This data provided by the API is updated hourly.
}
\examples{
uvi()
uvi(date = "2022-09-03T16:45:23")
uvi(date = "2021-12-06T11:01:55")
}
\keyword{UVI}
