% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVISummary.R
\name{uvi_summary}
\alias{uvi_summary}
\title{UVI(Ultra-violet Index) Summary}
\usage{
uvi_summary(date = "")
}
\arguments{
\item{date}{Defaults to current (SGD) time. Format: YYYY-MM-DD}
}
\value{
A dataframe containing the UVI for the given date
}
\description{
This functions calls upon the UVI API from data.gov.sg
and returns UVI values for the given date.
This data provided by the API is updated hourly.
}
\details{
Note that there is only information on the UVI between 0700 and 1900.
The returned data frame would only contain UVI values between this
range.
}
\examples{
uvi_summary()
uvi_summary(date = "2018-10-10")
uvi_summary(date = "2019-04-01")
}
\keyword{UVI}
