% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarparkAvailability.R
\name{carpark_availability}
\alias{carpark_availability}
\title{Carpark Availability}
\usage{
carpark_availability(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing the carpark id, type, last update, total lots, and current lots.
}
\description{
This functions calls upon the carpark availability API from data.gov.sg
and processes the returning page.
}
\examples{
\donttest{
carpark_availability()
carpark_availability(date = "2019-06-05T10:10:10")
carpark_availability(date = "2018-12-01T19:32:56")
}
}
\keyword{carpark}
