% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseApiOutput.R
\name{parse_api_output}
\alias{parse_api_output}
\title{ParseApiOutput}
\usage{
parse_api_output(inputcontent)
}
\arguments{
\item{inputcontent}{Takes the output of the GET function, runs an error check and returns the parsed output.}
}
\value{
The extracted content if not error has occured. Otherwise, the error message is returned.
}
\description{
Helper function to extract the content returned from the API.
Returns the status code otherwise.
}
