% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVI.R
\name{uvi}
\alias{uvi}
\title{UVI(Ultra-violet Index)}
\usage{
uvi(date_time = "")
}
\arguments{
\item{date_time}{Defaults to current (SGD) time. Format: YYYY-MM-DDTHH:MM:SS}
}
\value{
A dataframe containing the UVI for the given date-time
}
\description{
This functions calls upon the UVI API from data.gov.sg
and returns the UVI value for the given data-time.
This data provided by the API is updated hourly.
}
\details{
Note that there is only information on the UVI between 0700 and 1900.
Entering a date-time outside of the range would return the UVI of the
closest date time instead.
}
\examples{
uvi()
uvi(date_time = "2018-10-10T15:00:00")
uvi(date_time = "2019-04-01T09:00:00")
}
\keyword{UVI}
